3 commit

+++ src/main/java/Context.java
@@ -1,3 +1,4 @@
+import java.math.BigDecimal;
 
 public class Context {
     private Operation operation;
@@ -6,7 +7,7 @@ public class Context {
         this.operation = operation;
     }
 
-    public double executeStrategy(double num1, double num2){
+    public BigDecimal executeStrategy(BigDecimal num1, BigDecimal num2){
         return operation.doOperation(num1, num2);
     }
 }
