+++ src/main/java/Operation.java
@@ -1,4 +1,5 @@
+import java.math.BigDecimal;
 
 public interface Operation {
-    public double doOperation(double num1, double num2);
+    public BigDecimal doOperation(BigDecimal num1, BigDecimal num2);
 }
